Add delete instance on test file
