+++ pkg/testcase/cluster.go
@@ -3,9 +3,7 @@ package testcase
 import (
 	"fmt"
 	"strings"
-	"sync"
 
-	"github.com/rancher/distros-test-framework/pkg/aws"
 	"github.com/rancher/distros-test-framework/pkg/customflag"
 	"github.com/rancher/distros-test-framework/shared"
 
@@ -100,26 +98,3 @@ func checkAndPrintAgentNodeIPs(agentNum int, agentIPs []string, isWindows bool) 
 	}
 }
 
-// DeleteAWSInstance deletes ec2 instance via aws client.
-func DeleteAWSInstance(cluster *shared.Cluster) {
-	ips := shared.FetchNodeExternalIPs()
-	awsClient, err := aws.AddClient(cluster)
-	if err != nil {
-		shared.LogLevel("error", "error creating aws client: %v", err)
-	}
-
-	var wg sync.WaitGroup
-	for _, ip := range ips {
-		ip := ip
-		wg.Add(1)
-		go func(ip string) {
-			defer wg.Done()
-			nodeDelErr := awsClient.DeleteInstance(ip)
-			if nodeDelErr != nil {
-				shared.LogLevel("error", "on deleting node with ip: %v, got error %w", ip, nodeDelErr)
-				return
-			}
-		}(ip)
-	}
-	wg.Wait()
-}
